#34 Fix progress report and test of StanfordNLPModule
mockito is way easier than running it in a separate thread, and it makes
the result more consistent

+++ src/main/java/de/unistuttgart/vis/vita/analysis/modules/StanfordNLPModule.java
@@ -79,7 +79,7 @@ public class StanfordNLPModule implements Module<StanfordNLPResult> {
       public Object apply(Annotation in) {
         progressInChars += in.get(TextAnnotation.class).length();
         double progress = MODEL_LOAD_PROGRESS_FRACTION +
-            (progressInChars / totalLength) * (1 - MODEL_LOAD_PROGRESS_FRACTION);
+            ((double)progressInChars / totalLength) * (1 - MODEL_LOAD_PROGRESS_FRACTION);
         progressListener.observeProgress(progress);
         return null; // ignored
       }
